Split AlgoScreen into components, refactor bigo
